requirements.txt
apns2==0.7.1
asgiref==3.2.7
Babel==2.8.0
blessings==1.7
bpython==0.19
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
curtsies==0.3.1
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.41.0
django-debug-toolbar==2.2
django-debug-toolbar-force==0.1.8
django-elasticsearch-debug-toolbar==2.0.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-extensions==2.2.9
django-extra-views==0.11.0
django-filter==2.2.0
django-haystack==2.8.1
django-nine==0.2.3
django-oscar==2.0.4
django-oscar-accounts==2.0.0
django-oscar-api==2.0.2
django-oscar-api-checkout==1.0.0
django-phonenumber-field==2.0.1
django-push-notifications==2.0.0
django-redis==4.11.0
django-rest-auth==0.9.5
django-solo==1.1.3
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
factory-boy==2.12.0
Faker==4.0.3
future==0.18.2
greenlet==0.4.15
gunicorn==19.7.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
pgeocode==0.2.1
phonenumbers==8.12.1
Pillow==7.1.1
psycopg2-binary==2.8.5
purl==1.5
py-vapid==1.7.0
pycountry==19.8.18
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.11.0
razorpay==1.2.0
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
stripe==2.48.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
